feat: pages into dynamic locale route
