+++ wealth-sync/src/app/[locale]/auth/layout.tsx
