+++ wealth-sync/src/app/[locale]/auth/login/page.tsx
