+++ wealth-sync/src/app/[locale]/auth/register/page.tsx
