+++ wealth-sync/src/app/[locale]/coming-soon/layout.tsx
