+++ wealth-sync/src/app/[locale]/coming-soon/page.tsx
