+++ wealth-sync/src/app/[locale]/connect/coinbase/laoyout.tsx
