+++ wealth-sync/src/app/[locale]/connect/etoro/page.tsx
