+++ wealth-sync/src/app/[locale]/connect/layout.tsx
