+++ wealth-sync/src/app/[locale]/connect/meta-mask/page.tsx
@@ -3,23 +3,23 @@
 import {
   withApiConnection,
   type ApiConnectionInjectedProps,
-} from "@/app/hocs/withApiConnection";
+} from "@/app/[locale]/hocs/withApiConnection";
 import { Button } from "@/components/ui/button";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { Input } from "@/components/ui/input";
 import { PlatformKey } from "@/lib/constants/apiKeyStrings";
 import { Routes } from "@/lib/constants/routes";
 import { usePlatformConnection } from "@/lib/contexts/PlatformConnectionContext";
-import { AlignJustify } from "lucide-react";
 import Image from "next/image";
 import { useRouter } from "next/navigation";
 import { useState } from "react";
 import { toast } from "sonner";
 
-function Trading212ConnectPage({
+function MetaMaskConnectPage({
   apiKey,
   setApiKey,
 }: ApiConnectionInjectedProps) {
+  const [apiSecret, setApiSecret] = useState<string>("");
   const [isLoading, setIsLoading] = useState<boolean>(false);
   const router = useRouter();
   const { refreshConnections } = usePlatformConnection();
@@ -29,30 +29,26 @@ function Trading212ConnectPage({
     setIsLoading(true);
 
     try {
-      const response = await fetch("/api/platforms/trading212/connect", {
+      await fetch("/api/platforms/meta-mask/connect", {
         method: "POST",
         headers: {
           "Content-Type": "application/json",
         },
-        body: JSON.stringify({ apiKey }),
+        body: JSON.stringify({ apiKey, apiSecret }),
       });
 
-      const responseData = (await response.json()) as Record<string, string>;
-
-      if (responseData.error) {
-        throw Error(responseData.error);
-      }
-
       await refreshConnections();
-      toast.success("Successfully connected to Trading212");
+      toast.success("Successfully connected to MetaMask");
       router.push(Routes.DASHBOARD);
     } catch (error) {
       setApiKey("");
+      setApiSecret("");
       toast.error("Failed to connect", {
         description:
           error instanceof Error ? error.message : "Please try again",
       });
     } finally {
+      setApiKey("");
       setIsLoading(false);
     }
   };
@@ -60,23 +56,25 @@ function Trading212ConnectPage({
   return (
     <>
       <div className="text-center">
-        <h1 className="text-3xl font-bold">Connect Trading212</h1>
+        <h1 className="text-3xl font-bold">Connect Meta Mask Wallet</h1>
         <p className="text-muted-foreground mt-2">
-          Connect your Trading212 account to start tracking your investments
+          Connect your MetaMask account to track your crypto portfolio
         </p>
       </div>
 
       <Card>
         <CardHeader>
           <div className="flex items-center gap-4">
-            <Image
-              src="/platforms/trading-212-icon.png"
-              alt="Trading212"
-              width={48}
-              height={48}
-              className="rounded-full"
-            />
-            <CardTitle>Trading212 API Connection</CardTitle>
+            <div className="relative">
+              <Image
+                src="/platforms/meta-mask-logo.png"
+                alt="MetaMask"
+                width={50}
+                height={50}
+                className="rounded-full"
+              />
+            </div>
+            <CardTitle>MetaMask API Connection </CardTitle>
           </div>
         </CardHeader>
         <CardContent>
@@ -85,14 +83,23 @@ function Trading212ConnectPage({
               <label className="text-sm font-medium">API Key</label>
               <Input
                 type="password"
-                placeholder="Enter your Trading212 API key"
+                placeholder="Enter your MetaMask API key"
                 value={apiKey}
                 onChange={(e) => setApiKey(e.target.value)}
                 required
               />
+            </div>
+            <div className="space-y-2">
+              <label className="text-sm font-medium">API Secret</label>
+              <Input
+                type="password"
+                placeholder="Enter your MetaMask API secret"
+                value={apiSecret}
+                onChange={(e) => setApiSecret(e.target.value)}
+                required
+              />
               <p className="text-muted-foreground text-xs">
-                You can find your API key in Trading212 settings under the API
-                section
+                Make sure to use read-only API keys for security
               </p>
             </div>
             <Button
@@ -100,7 +107,7 @@ function Trading212ConnectPage({
               className="w-full cursor-pointer hover:bg-green-500"
               disabled={isLoading}
             >
-              {isLoading ? "Connecting..." : "Connect Trading212"}
+              {isLoading ? "Connecting..." : "Connect MetaMask"}
             </Button>
           </form>
         </CardContent>
@@ -108,27 +115,15 @@ function Trading212ConnectPage({
 
       <Card>
         <CardHeader>
-          <CardTitle>How to get your API key</CardTitle>
+          <CardTitle>How to get your API credentials</CardTitle>
         </CardHeader>
         <CardContent className="space-y-4">
           <ol className="list-decimal space-y-2 pl-4">
-            <li>
-              <a
-                href="https://app.trading212.com/"
-                target="_blank"
-                rel="noopener noreferrer"
-                className="hover:text-blue-400 hover:underline"
-              >
-                Log in to your Trading212 account
-              </a>
-            </li>
-            <div className="flex items-center gap-2">
-              <li>Go to Account Settings </li>
-              <AlignJustify className="h-4 w-4" />
-            </div>
-            <li>Navigate to the API section</li>
-            <li>Click on &quot;Generate new API key&quot;</li>
-            <li>Copy the generated key and paste it above</li>
+            <li>Log in to your MetaMask account</li>
+            <li>Go to API Management</li>
+            <li>Click &quot;Create API&quot;</li>
+            <li>Set permissions to &quot;Read Only&quot;</li>
+            <li>Copy both the API key and secret</li>
           </ol>
         </CardContent>
       </Card>
@@ -136,7 +131,4 @@ function Trading212ConnectPage({
   );
 }
 
-export default withApiConnection(
-  Trading212ConnectPage,
-  PlatformKey.TRADING_212,
-);
+export default withApiConnection(MetaMaskConnectPage, PlatformKey.META_MASK);
