+++ wealth-sync/src/app/[locale]/contact/page.tsx
