+++ wealth-sync/src/app/[locale]/dashboard/layout.tsx
