+++ wealth-sync/src/app/[locale]/dashboard/page.tsx
