+++ wealth-sync/src/app/[locale]/docs/page.tsx
