+++ wealth-sync/src/app/[locale]/features/page.tsx
