+++ wealth-sync/src/app/[locale]/hocs/withApiConnection.tsx
