+++ wealth-sync/src/app/[locale]/integrations/page.tsx
