+++ wealth-sync/src/app/[locale]/layout.tsx
