+++ wealth-sync/src/app/[locale]/loading.tsx
