+++ wealth-sync/src/app/[locale]/not-found.tsx
