+++ wealth-sync/src/app/[locale]/page.tsx
