+++ wealth-sync/src/app/[locale]/pricing/page.tsx
