+++ wealth-sync/src/app/[locale]/privacy/page.tsx
