+++ wealth-sync/src/app/[locale]/profile/layout.tsx
