+++ wealth-sync/src/app/[locale]/profile/page.tsx
