+++ wealth-sync/src/app/[locale]/settings/layout.tsx
