+++ wealth-sync/src/app/[locale]/settings/page.tsx
