+++ wealth-sync/src/app/[locale]/terms/page.tsx
